proteinweb/requirements.txt
aioredis==1.3.1
amqp==5.0.9
asgiref==3.4.1
asttokens==2.0.5
async-timeout==4.0.2
atomicwrites==1.4.0
attrs==21.4.0
autobahn==22.1.1
Automat==20.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
billiard==3.6.4.0
celery==5.0.5
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==2.0.12
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
constantly==15.1.0
cryptography==36.0.1
daphne==3.0.2
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.4
Django==3.0.3
django-allauth==0.48.0
django-bootstrap4==2.2.0
django-crispy-forms==1.14.0
djangorestframework==3.11.1
drf-writable-nested==0.6.3
executing==0.8.3
factory-boy==3.0.1
Faker==9.8.3
filelock==3.6.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
iniconfig==1.1.1
ipython==8.3.0
jedi==0.18.1
kombu==5.2.3
matplotlib-inline==0.1.3
msgpack==1.0.3
networkx==1.11
numpy==1.17.3
oauthlib==3.2.0
packaging==21.3
parso==0.8.3
pickleshare==0.7.5
Pillow==7.2.0
platformdirs==2.5.1
pluggy==1.0.0
ply==3.11
prompt-toolkit==3.0.26
psycopg2==2.8.6
pure-eval==0.2.2
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.12.0
PyJWT==2.3.0
pyOpenSSL==22.0.0
pyparsing==3.0.7
pytest==7.1.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
redis==3.5.3
requests==2.27.1
requests-oauthlib==1.3.1
scipy==1.8.0
service-identity==21.1.0
six==1.16.0
smop==0.41b0
soupsieve==2.3.1
sqlparse==0.4.2
stack-data==0.2.0
text-unidecode==1.3
tomli==2.0.1
traitlets==5.2.0
Twisted==22.1.0
twisted-iocpsupport==1.0.2
txaio==21.2.1
typing-extensions==4.1.1
tzdata==2021.5
unrar==0.4
urllib3==1.26.8
vine==5.0.0
virtualenv==20.13.1
wcwidth==0.2.5
wrapt==1.13.3
zope.interface==5.4.0
